ADD #63 - modal 컴포넌트 분리
