+++ client/src/ProfilePageView/UserProfile.js
@@ -1,15 +1,12 @@
 import { useEffect, useRef, useState } from 'react';
 import { MdFormatListNumberedRtl, MdModeEdit } from 'react-icons/md';
-import { FiCheck } from 'react-icons/fi';
-import { IoClose } from 'react-icons/io5';
 import axios from 'axios';
+import EditNickModal from './EditNickModal';
 
 function UserProfile() {
   const [userImage, setUserImage] = useState('');
   const [userNickname, setUserNickname] = useState('');
-  const [newNickname, setNewNickname] = useState('');
   const [editNickModal, setEditNickModal] = useState(false);
-  const modalRef = useRef();
 
   const defaultUserImage =
     'https://karateinthewoodlands.com/wp-content/uploads/2017/09/default-user-image-300x300.png';
@@ -29,20 +26,11 @@ function UserProfile() {
     // user image 요청 추가
   }, [userImage]);
 
-  const onChangeNewNick = e => {
-    setNewNickname(e.target.value);
-  };
-
   const editUserNickname = () => {
     console.log(`edit btn click!!!`);
     setEditNickModal(!editNickModal);
     console.log(editNickModal);
   };
-
-  window.addEventListener('click', e => {
-    if (e.target === modalRef.current) editUserNickname();
-    return false;
-  });
 
   return (
     <>
@@ -60,33 +48,7 @@ function UserProfile() {
           </div>
         </div>
       </section>
-
-      {editNickModal && (
-        <div className="modal-container" ref={modalRef}>
-          <div className="edit-nick-modal">
-            <IoClose
-              className="edit-nick-modal-close-btn"
-              onClick={editUserNickname}
-            />
-            <span className="edit-nick-modal__title">
-              새 닉네임을 입력하세요
-            </span>
-            <div className="edit-nick-modal__input-box">
-              <input
-                type="text"
-                className="edit-nick-modal__input"
-                onChange={onChangeNewNick}
-                autoFocus
-              />
-              <FiCheck className="edit-nick-modal__input-check-btn" />
-            </div>
-            <span className="edit-nick-modal__check-msg"></span>
-            <button className="edit-nick-modal__input-submit-btn">
-              저장하기
-            </button>
-          </div>
-        </div>
-      )}
+      {editNickModal && <EditNickModal editUserNickname={editUserNickname} />}
     </>
   );
 }
